Use NSOperationQueue to single-thread access to TDServer/Database.
We no longer run everything on the NSURLConnection thread. Instead the TDServer creates an NSOperationQueue and everything else that talks to the server or its database does so via blocks posted to the queue.
This required changing a bunch of code (URL loading, stream, delayed perform) that depended on runloops to use operation or dispatch queues instead.