Code coverage: check for methods not called during their tests
Now checks whether a Cover() call isn’t reached at all during the test
case that claims to test it (the one in the TestedBy() declaration.)
This is tested at the end of the test run and treated as a special test
case.
Also fixed the MYURLUtils tests to no longer fail coverage checks.