MYDynamicObject: Workaround for handling Swift-class-valued property types
If a property type is a Swift class, the runtime type metadata doesn't
include the class's namespace, which causes the class lookup to fail.
Worked around this by a heuristic of trying the namespace of the class
that contains the property. (The underlying problem is Apple's; the
Radar ID is 21368142.)