Added accessible param to MYGetOrCreateAnonymousIdentity
Added accessible param whose value indicates when a keychain item is accessible. If passing a NULL value, the kSecAttrAccessible key will not be set to the keys/certificates when saving them in the keychain; as a result, the default kSecAttrAccessible’s value from the OS which is kSecAttrAccessibleWhenUnlocked will be automatically applied.

Reference: https://github.com/couchbase/couchbase-lite-ios/issues/2165